refactor: 컴포넌트를 분리한다
